+++ test/src/util/string_extension_test.dart
@@ -1,4 +1,4 @@
-import 'package:calendar_timeline/src/util/string_extension.dart';
+import 'package:calendar_timeline_sbk/src/util/string_extension.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import '../../test_utils/param_factory.dart';